requirements.txt
pandas>=2.3.3
openpyxl>=3.1.5
watchdog>=6.0
python-dotenv>=1.2.1
psycopg>=3.3.2